initial quantity move
