+++ astropy/quantity/__init__.py
@@ -1,0 +1,11 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+"""
+This subpackage contains the `Quantity` class and associated tools.
+It provides an easy but powerful way to represents scalar or array
+values with units. `Quantity` objects support operations like ordinary
+numbers, but deal with unit conversions internally.
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from .core import *
